requirements.txt
pandas
plotly
streamlit
python-dotenv
anthropic